implement feedback and move test_only_s3_targets_expires_header() to test_s3.py
